[gridworld] lib value state_action_value action_value
