+++ gridworld/src/mdp.rs
@@ -4,6 +4,8 @@ mod environment;
 mod policy;
 mod solvers;
 mod state;
+mod transition;
+mod value;
 
 pub use action::*;
 pub use action_value_map::*;
@@ -11,3 +13,5 @@ pub use environment::*;
 pub use policy::*;
 pub use solvers::*;
 pub use state::*;
+pub use transition::*;
+pub use value::*;
