+++ gridworld/src/mdp/solvers.rs
@@ -2,6 +2,7 @@ use super::{Environment, Policy};
 
 /// Evaluates the policy for a given environment and optional initial state value grid
 /// Returns a new state value grid
+fn test() {}
 
 // pub fn policy_evaluation(
 //     cell_grid: &Grid<Cell>,
