+++ gridworld/src/mdp/state.rs
@@ -1,1 +1,5 @@
-pub trait State {}
+use std::hash::Hash;
+
+pub trait State: Sized + Hash + Eq {
+    fn get_all() -> Vec<Self>;
+}
