+++ gridworld/src/mdp/transition.rs
@@ -1,0 +1,1 @@
+pub type Transition<S, A> = fn(&S, &A) -> (S, f64);
